requirements.txt
numpy
Pillow
opencv-python
tifffile